move all deprecated methods to deprecated.kt
